src/shared/domain/Result.ts
export class Result<T> {

  public isSuccess: boolean;
  public error: string;
  private readonly value: T;

  public constructor(isSuccess: boolean, error?: string, value?: T) {
    this.isSuccess = isSuccess;
    this.error = error;
    this.value = value;

    Object.freeze(this);
  }

  public get getValue(): T {
    if (!this.isSuccess) {
      throw new Error("Can't get the value of an error result. Use 'errorValue' instead.");
    }

    return this.value;
  }

  public static ok<U>(value?: U): Result<U> {
    return new Result<U>(true, null, value);
  }

  public static fail<U>(error: string): Result<U> {
    return new Result<U>(false, error);
  }
}
